+++ charts/collector-clair/templates/service.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: {{ include "collector-clair.fullname" . }}
+  labels:
+    {{- include "collector-clair.labels" . | nindent 4 }}
+spec:
+  type: {{ .Values.service.type }}
+  ports:
+    - port: {{ .Values.service.port }}
+      targetPort: http
+      protocol: TCP
+      name: http
+  selector:
+    {{- include "collector-clair.selectorLabels" . | nindent 4 }}
