+++ charts/collector-clair/templates/tests/test-connection.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: "{{ include "collector-clair.fullname" . }}-test-connection"
+  labels:
+    {{- include "collector-clair.labels" . | nindent 4 }}
+  annotations:
+    "helm.sh/hook": test-success
+spec:
+  containers:
+    - name: wget
+      image: busybox
+      command: ['wget']
+      args: ['{{ include "collector-clair.fullname" . }}:{{ .Values.service.port }}']
+  restartPolicy: Never
